+++ src/utils.test.js
@@ -14,10 +14,12 @@ import {
   getInstanceQueryString,
   generateUserName,
   handleKeyCommand,
+  isValidRequest,
   memoizeValidation,
 } from './utils';
 
 import {
+  INVALID_REQUEST_HARDCODED_ID,
   itemStatuses,
   requestTypesMap,
   REQUEST_LEVEL_TYPES,
@@ -313,6 +315,44 @@ describe('handlekeycommand', () => {
     handleKeyCommand(handler)();
     expect(handler).toHaveBeenCalled();
     expect(event.preventDefault).not.toHaveBeenCalled();
+  });
+});
+
+describe('isValidRequest', () => {
+  it('should return true if request is valid', () => {
+    const request = {
+      instanceId: 'testInstanceId',
+      holdingsRecordId: 'testHoldingRecordId',
+    };
+
+    expect(isValidRequest(request)).toBe(true);
+  });
+
+  it('should return false if "instanceId" in request have hardcoded invalid value', () => {
+    const request = {
+      instanceId: INVALID_REQUEST_HARDCODED_ID,
+      holdingsRecordId: 'testHoldingRecordId',
+    };
+
+    expect(isValidRequest(request)).toBe(false);
+  });
+
+  it('should return false if "holdingsRecordId" in request have hardcoded invalid value', () => {
+    const request = {
+      instanceId: 'testInstanceId',
+      holdingsRecordId: INVALID_REQUEST_HARDCODED_ID,
+    };
+
+    expect(isValidRequest(request)).toBe(false);
+  });
+
+  it('should return false if "instanceId" and "holdingsRecordId" in request have hardcoded invalid value', () => {
+    const request = {
+      instanceId: INVALID_REQUEST_HARDCODED_ID,
+      holdingsRecordId: INVALID_REQUEST_HARDCODED_ID,
+    };
+
+    expect(isValidRequest(request)).toBe(false);
   });
 });