+++ test/jest/__mock__/stripesSmartComponents.mock.js
@@ -1,6 +1,7 @@
 import React from 'react';
 
 jest.mock('@folio/stripes/smart-components', () => ({
+  ClipCopy: jest.fn(() => null),
   makeQueryFunction: jest.fn((value) => value),
   CheckboxFilter: jest.fn(() => null),
   NotesSmartAccordion: jest.fn(() => null),